ci: Don't run placeholder instrumenation tests (#119)
Due to a bug in `mobile-android-pipelines` Jacoco and Sonar plugins it isn't possible to create an Android library without an associated instrumentation test. Hence we have a lot of tests that add no value.

Until we can delete these tests, this update the workflow so that they aren't run at all.